msm: mdss: dsi: fix clk_rate to use u64 instead of u32
Commit f9c04f165f9bb98fe7dbc4c605fba7f34bc4e23c ("msm: mdss: dsi: add
support for multiple panel timing settings") changed the
mdss-dsi-panel-clockrate parameter read from u64 to u32 during the
restructuring to use panel timings struct. Revert it back to use u64.

+++ drivers/video/msm/mdss/mdss_panel.h
@@ -655,7 +655,7 @@ struct mdss_panel_timing {
 
 	u32 lm_widths[2];
 
-	u32 clk_rate;
+	u64 clk_rate;
 	char frame_rate;
 
 	u8 dsc_enc_total;
